backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2022.5
sentry-sdk==1.10.1
sqlparse==0.4.3
urllib3==1.26.12
